+++ utils.go
@@ -3,6 +3,7 @@ package diskq
 import (
 	"encoding/binary"
 	"fmt"
+	"hash/fnv"
 	"io"
 	"io/fs"
 	"os"
@@ -310,4 +311,19 @@ func GetPartitionSegmentStartOffsets(path string, partitionIndex uint32) (output
 		output = append(output, segmentStartOffset)
 	}
 	return
+}
+
+// internal utils
+
+func hashIndexForMessage(m Message, partitions int) int {
+	h := fnv.New32()
+	_, _ = h.Write([]byte(m.PartitionKey))
+	return int(h.Sum32()) % partitions
+}
+
+func maybeSync(wr io.Writer) error {
+	if typed, ok := wr.(*os.File); ok {
+		return typed.Sync()
+	}
+	return nil
 }
